Add files via upload
This file contains queries that create tables and also insert values into these tables for analysis purposes. It mainly contains the CREATE statement and the INSERT statement; along with data types like INTEGER for numeric values, VARCHAR and TEXT for letters as well as DATETIME for date values. Statements and datetypes are all in capital letters, while values, table names and column names are all in lower cases.